AppClientIntro- with TODOs
